유저 비밀번호 재설정시 권한 문제 해결, 회원 탈퇴 기능 구현 중

+++ trip-location/src/main/java/com/korea/triplocation/service/UserService.java
@@ -17,6 +17,7 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import com.korea.triplocation.api.dto.request.LoginReqDto;
 import com.korea.triplocation.api.dto.request.ResetPasswordReqDto;
 import com.korea.triplocation.api.dto.request.UserModifyReqDto;
 import com.korea.triplocation.api.dto.response.UserRespDto;
@@ -196,8 +197,10 @@ public class UserService {
 		return userRepository.resetPassword(user) != 0;
 	}
 	
-	public void deleteUser(int userId) {
-		userRepository.deleteUser(userId);
+	public void deleteUser(int userId, LoginReqDto loginReqDto) {
+		if(authService.signin(loginReqDto) != null) {
+			userRepository.deleteUser(userId);			
+		}
 	}
 
 }
